增加两个hmap debug设置用的api
**复现步骤**

无

**原因和解决方案**

方便设置参数

**影响范围**

无

**验证方案**

ut通过

**涉及分支**

* master

**检查项**

- [ ] 需要更新依赖
- [ ] 是共性问题(代码中存在类似问题)
- [ ] 编译通过
- [ ] 单元测试通过

/assign @xiangyang

/cc @yuanchao @jinjie @zhuofeng @wangtuo

+++ hmap/idmap/u128_id_map.go
@@ -157,8 +157,8 @@ func (m *U128IDMap) GetCollisionChain() []byte {
 	return chain
 }
 
-func (m *U128IDMap) SetCollisionChainDebugThreshold(t uint32) {
-	atomic.StoreUint32(&m.collisionChainDebugThreshold, t)
+func (m *U128IDMap) SetCollisionChainDebugThreshold(t int) {
+	atomic.StoreUint32(&m.collisionChainDebugThreshold, uint32(t))
 }
 
 // 第一个返回值表示value，第二个返回值表示是否进行了Add。若key已存在，指定overwrite=true可覆写value。
